Converted string dates to datetime type
The water_data table formerly had dates stored as a "string" type. A
script was created to convert the "string" dates to "datetime" type
dates and run in the rtidev database. This script will me altered
slightly an run in production after altering the Database-Updater and
RTI Web applications for compatibility.